libvirt: Add configuration of guest VCPU topology
Wire up the libvirt driver to honour the VCPU topology config
requested via the flavor extra specs / image meta. This is
only done for QEMU/KVM guests at this time. Although Xen is
technically capable of supporting this, it is not yet wired
up in libvirt itself.

Blueprint: virt-driver-vcpu-topology
Change-Id: I4540147af7b91e1dc463be63d19afbfcfc1216cb

Conflicts:
	nova/tests/virt/libvirt/test_libvirt.py
	nova/virt/libvirt/driver.py